Several enhancements to add better indexing of words. Now all unique wordforms are added to the index, meaning that all these forms are recognised as variants of the word, and will automatically display the relevant dictionary article.
NB! Because of the way this works, words containing non-ASCII chars are presently NOT included. This is most likely because the URI constructed in XSLT is incompatible with the underlying pathname in OS X (precomposed versus decomposed Unicode chars). This should really be handled transparently by the OS, but it is not.

Soo.... not all words will have paradigmes - yet. I'm working on it.